(NOTAG) initial commit for `maven-smartmodules-extension`
Basic support for automatic discovery of descendant modules.

Beware: not thoroughly tested, not verified for consistency
of the internal Maven Model of the project tree, and may
not respect the order of the projects as they appear in
the POMs if we no other rules apply.

Packaged with a basic example and integration test, which
contains a multi-module build tree with several levels of
descendant modules, but:

 - *without* any inter-module dependencies,
 - *without* any build order assumptions,
 - *without* any POM inheritance.